Monolith split complete — app shell moved to core/shell.py, main file down to 625 lines.
Completes the split. The residual bootstrap — creating the Tk root, the Notebook
scaffold, tab initialisation and the main loop — is relocated into core/shell.py
(822 lines). lora_trainer_gui.py now holds only the entry point and a thin import
list. Monolith: 1,456 → 625 lines (from 27,640 at the start of the branch — −97.7%).
Pure move; behaviour is bit-for-bit identical to master. The split is finished;
per-module refactoring of each module (decoupling from self, removing forward
references, unit tests) is now unblocked and lands separately — one module at a
time.
